Move the error handler to a new file
This will allow it to be triggered earlier in the command lifecycle
